Always allow 15 minutes of action time
This changes the post timer to always allow 15 minutes of action
time, regardless of expected state. This way, if somebody accepts the
 post too late, they still get 15 minutes before they get the
 reminder.

Further, this removes duplicated code that used to directly switch
from action to preparation - this is now done via a temporary switch
to idle.

Lastly, it fixes the issue where no card reminder would be displayed
if the expected state was already idle in action state.